+++ FitpaySDK/Rest/Models/EncryptionKey.swift
@@ -4,7 +4,7 @@ public class EncryptionKey
     public var links:[ResourceLink]?
     public var keyId:String?
     public var created:String?
-    public var createdEpoch:Int?
+    public var createdEpoch:Int64?
     public var serverPublicKey:String?
     public var clientPublicKey:String?
 }
